refactor: change replaceGCSUrlOrigin to a utility function
