+++ fields/types/gcsfile/GcsFileField.js
@@ -1,22 +1,8 @@
 import Field from '../Field';
 import React from 'react';
 import ReactDOM from 'react-dom';
+import replaceGCSUrlOrigin from '../../utils/replace-gcs-url-origin';
 import { Button, FormField, FormInput, FormNote } from 'elemental';
-
-const storage = {
-  google: {
-    schema: 'https',
-    hostname: 'storage.googleapis.com',
-    bucket: 'twreporter-multimedia',
-  },
-};
-
-function replaceGCSUrlOrigin(url = '') {
-  const { schema, hostname, bucket } = storage.google;
-  const toReplace = 'https://www.twreporter.org';
-  const toBeReplaced = `${schema}://${hostname}/${bucket}`;
-  return url.replace(toBeReplaced, toReplace);
-}
 
 module.exports = Field.create({
