+++ fields/types/gcsimage/GcsImageField.js
@@ -3,22 +3,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import get from 'lodash/get';
 import keys from 'lodash/keys';
+import replaceGCSUrlOrigin from '../../utils/replace-gcs-url-origin';
 import { Button, FormField, FormInput, FormNote } from 'elemental';
-
-const storage = {
-  google: {
-    schema: 'https',
-    hostname: 'storage.googleapis.com',
-    bucket: 'twreporter-multimedia',
-  },
-};
-
-function replaceGCSUrlOrigin(url = '') {
-  const { schema, hostname, bucket } = storage.google;
-  const toReplace = 'https://www.twreporter.org';
-  const toBeReplaced = `${schema}://${hostname}/${bucket}`;
-  return url.replace(toBeReplaced, toReplace);
-}
 
 const _ = {
   get,
